7 ajax

+++ app/controllers/answers_controller.rb
@@ -5,30 +5,21 @@ class AnswersController < ApplicationController
   expose :question, -> { Question.find(params[:question_id]) }
 
   def create
-    @exposed_answer = question.answers.new(answer_params.merge( user_id: current_user.id))
-
-    if answer.save
-      redirect_to question, notice: 'Your answer successfully created.'
-    else
-      render 'questions/show'
-    end
+    @exposed_answer = question.answers.create(answer_params.merge( user_id: current_user.id))
   end
 
   def update
-    if answer.update(answer_params)
-      redirect_to answer_path(answer)
-    else
-      render :edit
-    end
+    answer.update(answer_params) if current_user.author?(answer)
+    @exposed_question = answer.question
   end
 
   def destroy
-    if current_user.author?(answer)
-      answer.destroy
-      redirect_to question_path(answer.question), notice: 'Answer successfully deleted.'
-    else
-      redirect_to question_path(answer.question), notice: 'You have no rigths to delete this answer.'
-    end
+    answer.destroy if current_user.author?(answer)
+  end
+
+  def mark_best
+    answer.mark_best if current_user.author?(answer.question)
+    @exposed_question = answer.question
   end
 
   private
